rtsp-publisher/requirements.txt
aiortc==1.9.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
python-socketio==5.11.2
aiohttp==3.8.6